+++ example/lib/main.dart
@@ -88,6 +88,17 @@ class _HomeState extends State<Home> {
             lastIconHighlight: true,
             lastIconHighlightColor: Colors.deepOrange,
             lastIconHighlightRadius: const Radius.circular(15.0),
+            captionTextStyle: TextStyle(
+              fontSize: 22,
+              color: Colors.white,
+            ),
+            captionMargin: EdgeInsets.only(
+              bottom: 50,
+            ),
+            captionPadding: EdgeInsets.symmetric(
+              horizontal: 24,
+              vertical: 8,
+            ),
           ),
           Center(
             child: Padding(